reference as issued_time rather than issuance and change percents to ints
