+++ src/ocha_lens/__init__.py
@@ -1,4 +1,4 @@
 from ._version import version as __version__  # noqa: F401
-from .datasources import ecmwf_storm, ibtracs, nhc, nhc_wsp
+from .datasources import ecmwf_storm, ibtracs, nhc
 
-__all__ = ["ibtracs", "ecmwf_storm", "nhc", "nhc_wsp"]
+__all__ = ["ibtracs", "ecmwf_storm", "nhc"]
